TODO: Clean UP

+++ HW2/BallsBlue.java
@@ -5,7 +5,7 @@ public class BallsBlue extends Decorations {
     public String printtree() {
         return tree.printtree() + ", Blue Ball";
     }
-    public int Cost(){
+    public double Cost(){
         return tree.Cost() + 2;
     }
