+++ HW2/BallsSilver.java
@@ -5,7 +5,7 @@ public class BallsSilver extends Decorations {
     public String printtree() {
         return tree.printtree() + ", Silver Ball";
     }
-    public int Cost(){
+    public double Cost(){
         return tree.Cost() + 3;
     }
